Too much stuff to list

+++ gE3/Engine/Asset/Buffer.cs
@@ -6,21 +6,21 @@ namespace gE3.Engine.Asset;
 public class Buffer : Asset
 {
     private readonly Target _target;
-    public int Size { get; }
+    public uint Size { get; }
     public uint Location { get; private set; }
-    public unsafe Buffer(GameWindow window, int size,
+    public unsafe Buffer(GameWindow window, uint size,
         Target target = Target.UniformBuffer) : base(window)
     {
         _target = target;
         Size = size;
         _id = GL.CreateBuffer();
-        GL.NamedBufferStorage(ID, (nuint)size, (void*)0, BufferStorageMask.DynamicStorageBit);
+        GL.NamedBufferStorage(ID, size, (void*)0, BufferStorageMask.DynamicStorageBit);
     }
     
-    public unsafe void ReplaceData(void* data, int size = 0, int offset = 0)
+    public unsafe void ReplaceData(void* data, uint size = 0, int offset = 0)
     {
         if (size == 0) size = Size;
-        GL.NamedBufferSubData(ID, offset, (nuint) size, data);
+        GL.NamedBufferSubData(ID, offset, size, data);
     }
     
     public void Bind(uint slot, Target? targetOverride = null)
